Address noise-floor PR review feedback: move RSSI sampling out of radio hot paths, use the public LR11x0 instant RSSI API, reject invalid -128 readings without clamping, and synchronize noise-floor telemetry state.

+++ src/mesh/aes-ccm.cpp
@@ -33,7 +33,7 @@ static int constant_time_compare(const void *a_, const void *b_, size_t len)
         d |= (a[i] ^ b[i]);
     }
     /* Constant time bit arithmetic to convert d > 0 to -1 and d = 0 to 0. */
-    return (1 & ((d - 1) >> 8)) - 1;
+    return (1 & (((unsigned int)d - 1) >> 8)) - 1;
 }
 
 static void WPA_PUT_BE16(uint8_t *a, uint16_t val)
